# docs/swagger.yaml
basePath: /api/v1
definitions:
  types.EntityRequest:
    properties:
      description:
        type: string
    type: object
  types.FailureErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  types.FailureResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  types.SuccessResponseData:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: This is an API of payment-service for Socialweb application
  title: Local-Template-Entity Swagger
  version: "1.0"
paths:
  /entity/create:
    post:
      consumes:
      - application/json
      description: Test endpoint crud
      parameters:
      - description: Entity data
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/types.EntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.SuccessResponseData'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: CreateEntity info
      tags:
      - Entity
  /entity/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Test endpoint crud
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.SuccessResponseData'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteEntity info
      tags:
      - Entity
  /entity/get:
    get:
      consumes:
      - application/json
      description: Test endpoint crud
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.SuccessResponseData'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetEntity info
      tags:
      - Entity
  /entity/test_grpc:
    get:
      consumes:
      - application/json
      description: Test endpoint for gRPC communication with auth service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SuccessResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Test gRPC connection to auth service
      tags:
      - grpc
  /entity/update/{id}:
    patch:
      consumes:
      - application/json
      description: Test endpoint crud
      parameters:
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Entity data
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/types.EntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.SuccessResponseData'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: UpdateEntity info
      tags:
      - Entity
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"